Update tpm clone to be quiet and prevent rm error

+++ home-manager/home.nix
@@ -26,8 +26,8 @@ in
 
         activation = {
             tpm = config.lib.dag.entryAfter ["writeBoundary"] ''
-                rm -rf ~/.tmux/plugins/tpm
-                ${pkgs.git}/bin/git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+                [[ -d ~/.tmux-plugins/tpm ]] && rm -rf ~/.tmux/plugins/tpm
+                ${pkgs.git}/bin/git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm &> /dev/null 2>&1
             '';
         };
